fix: app cpu quota type 수정

+++ python/mlad/service/models/app.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 
 class Quota(BaseModel):
-    cpu: Optional[int] = None
+    cpu: Optional[float] = None
     gpu: Optional[int] = None
     mem: Optional[str] = None
